- Updated tests - Moved out initial data loading

+++ src/main/java/com/embl/inventory/entity/ItemEntity.java
@@ -26,7 +26,7 @@ public class ItemEntity {
     @Column(unique = true)
     private String sku;
     @Column
-    private BigInteger count;
+    private int count;
 
     @Override
     public String toString() {
